+++ source/lsp/src/lib.rs
@@ -12,7 +12,7 @@ use std::{collections::hash_map::Entry, path::Path};
 
 pub use async_lsp;
 use async_lsp::{lsp_types as lsp, ClientSocket, LanguageClient};
-use flams_ontology::uris::{ArchiveId, DocumentURI};
+use flams_ontology::uris::{ArchiveId, BaseURI, DocumentURI};
 use flams_stex::quickparse::stex::{
     structs::{GetModuleError, ModuleReference, STeXModuleStore},
     STeXParseData,
@@ -169,6 +169,17 @@ impl lsp::notification::Notification for InstallArchives {
 }
 
 #[derive(serde::Serialize, serde::Deserialize)]
+struct NewArchiveParams {
+    pub archive: ArchiveId,
+    pub urlbase: BaseURI,
+}
+struct NewArchive;
+impl lsp::notification::Notification for NewArchive {
+    type Params = NewArchiveParams;
+    const METHOD: &str = "flams/newArchive";
+}
+
+#[derive(serde::Serialize, serde::Deserialize)]
 pub struct HtmlRequestParams {
     pub uri: lsp::Url,
 }
@@ -207,6 +218,12 @@ impl lsp::notification::Notification for UpdateMathHub {
     const METHOD: &str = "flams/updateMathHub";
 }
 
+struct OpenFile;
+impl lsp::notification::Notification for OpenFile {
+    type Params = lsp::Url;
+    const METHOD: &str = "flams/openFile";
+}
+
 struct HTMLResult;
 impl lsp::notification::Notification for HTMLResult {
     type Params = String;
@@ -228,6 +245,7 @@ impl lsp::notification::Notification for ServerURL {
 pub trait ClientExt {
     fn html_result(&self, uri: &DocumentURI);
     fn update_mathhub(&self);
+    fn open_file(&self, path: &Path);
 }
 impl ClientExt for ClientSocket {
     #[inline]
@@ -237,6 +255,16 @@ impl ClientExt for ClientSocket {
     #[inline]
     fn update_mathhub(&self) {
         if let Err(e) = self.notify::<UpdateMathHub>(()) {
+            tracing::error!("failed to send notification: {}", e);
+            return;
+        }
+    }
+
+    fn open_file(&self, path: &Path) {
+        let Ok(url) = lsp::Url::from_file_path(path) else {
+            return;
+        };
+        if let Err(e) = self.notify::<OpenFile>(url) {
             tracing::error!("failed to send notification: {}", e);
             return;
         }
@@ -273,6 +301,7 @@ impl<T: FLAMSLSPServer> ServerWrapper<T> {
         r.notification::<Reload>(Self::reload);
         r.notification::<StandaloneExport>(Self::export_standalone);
         r.notification::<InstallArchives>(Self::install);
+        r.notification::<NewArchive>(Self::new_archive);
         //r.request(handler)
         r
     }